beyond_challenge/requirements.txt
asgiref==3.8.1
astroid==3.2.4
asttokens==2.4.1
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dill==0.3.8
Django==5.1
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
executing==2.0.1
freezegun==1.5.1
idna==3.7
iniconfig==2.0.0
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pylint==3.2.6
pytest==8.3.2
pytest-django==4.8.0
pytest-freezegun==0.4.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
requests-mock==1.12.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomlkit==0.13.0
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.44.0
